Add configuration options
The Benchmark.ips(5,2) syntax is vague and difficult to understand.
While keeping this syntax in place, add an additional ability to
configure these options within the block itself by calling `config` on
the underlying IPSJob object and pass in a hash with :time and :warmup
symbols. Also add this to README and example script.